+++ src/lib/djerba/util/validator.py
@@ -11,7 +11,6 @@ class config_validator(logger):
     def __init__(self, wgs_only, failed, log_level=logging.WARNING, log_path=None):
         self.logger = self.get_logger(log_level, __name__, log_path)
         self.wgs_only = wgs_only
-        self.failed = failed
         if failed:
             self.logger.info("Validating config for failed report; requirements are the same for WGS-only and WGTS")
             self.schema = ini.SCHEMA_FAILED